rename: user, global 디렉토리 생성 및 이동
