+++ src/main/kotlin/org/scaleadvisor/backend/global/exception/model/UnauthorizedException.kt
@@ -1,6 +1,6 @@
-package org.scaleadvisor.backend.exception.model
+package org.scaleadvisor.backend.global.exception.model
 
-import org.scaleadvisor.backend.exception.BaseErrorCode
+import org.scaleadvisor.backend.global.exception.BaseErrorCode
 
 class UnauthorizedException(
     message: String,
