+++ src/test/kotlin/org/scaleadvisor/backend/JooqTest.kt
@@ -2,7 +2,7 @@ package org.scaleadvisor.backend
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
-import org.scaleadvisor.backend.user.UserRepository
+import org.scaleadvisor.backend.user.repository.UserRepository
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.transaction.annotation.Transactional
@@ -45,4 +45,6 @@ class JooqTest {
         assertThat(user).isNotNull
         assertThat(user!!.id).isEqualTo(userId)
     }
+
+
 }
